macos_menubar_app/requirements.txt
rumps>=0.4.0
pyobjc-framework-Cocoa>=10.0
pyobjc-framework-ApplicationServices>=10.0
pyobjc-framework-Quartz>=10.0
sounddevice>=0.4.6
numpy>=1.24.0
openai>=1.0.0
pyperclip>=1.8.2
scipy>=1.11.0
pynput>=1.7.6
